Dewarning on windows.
git-svn-id: svn://falconpl.org/falcon/core/trunk@540 78e5f83a-d9f7-4e86-b854-dad15b5eb37a

+++ engine/fasm_comp.cpp
@@ -160,7 +160,7 @@ const String *AsmCompiler::addString( const String &data )
 
 void AsmCompiler::addDLine( Pseudo *line )
 {
-   m_currentLine = line->asInt();
+   m_currentLine = (uint32) line->asInt();
    m_module->addLineInfo( m_pc + (uint32) m_outTemp->tell(), static_cast<uint32>(line->asInt()) );
    delete line;
 }
